pyproject.toml
[project]
name = "findingmodels"
version = "0.1.0"
description = "Finding models are updatable data models defining open, standard semantic tags for imaging findings and their properties"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonschema",
    "pydantic-ai",
    "python-dotenv",
    "pre-commit",
    "prettier>=0.0.7",
    "findingmodel>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "findingmodels", "scripts"]

[dependency-groups]
dev = [
    "ipython>=9.6.0",
]

